implementation/pam-system/gateway/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
paramiko==3.4.0
pyjwt==2.8.0
requests==2.31.0
